[pr1308] Posilani SetValue u timeru rac 64bitovy at se tam da treba v customValue ulozit vic veci

+++ src/game/SpellTimer.cpp
@@ -50,7 +50,7 @@ void SpellTimer::Reset(TimerValues value)
     }
 }
 
-void SpellTimer::SetValue(TimerValues value, uint32 newValue)
+void SpellTimer::SetValue(TimerValues value, uint64 newValue)
 {
     switch (value)
     {
